Update on "refactor ps benchmark"
This pr refactors the ps benchmark for modular trainers.

[ghstack-poisoned]
